# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "월계온 모바일 API",
    "version": "0.1.0",
    "description": "공사 우회·조명 기반 밤길 비교. 시연 데이터는 isDemo=true. GeoJSON 좌표는 [lng,lat]."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/auth/signup": {
      "post": {
        "summary": "월계동 주민 회원가입",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "required": [
                  "email",
                  "password",
                  "nickname",
                  "district"
                ],
                "properties": {
                  "email": {
                    "format": "email",
                    "maxLength": 254,
                    "type": "string"
                  },
                  "password": {
                    "minLength": 10,
                    "maxLength": 128,
                    "type": "string"
                  },
                  "nickname": {
                    "minLength": 1,
                    "maxLength": 30,
                    "type": "string"
                  },
                  "district": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "월계1동"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "월계2동"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "월계3동"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "월계동 생활권"
                        ]
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "summary": "주민 로그인",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "format": "email",
                    "maxLength": 254,
                    "type": "string"
                  },
                  "password": {
                    "minLength": 10,
                    "maxLength": 128,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "summary": "현재 로그인 세션 종료",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/v1/me": {
      "get": {
        "summary": "검증된 로그인 사용자와 운영 권한",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/coverage/nearby": {
      "get": {
        "summary": "인근 지도 연결 지점 후보. 자동 도로 연결 없음",
        "parameters": [
          {
            "schema": {
              "minimum": -90,
              "maximum": 90,
              "type": "number"
            },
            "in": "query",
            "name": "lat",
            "required": true
          },
          {
            "schema": {
              "minimum": -180,
              "maximum": 180,
              "type": "number"
            },
            "in": "query",
            "name": "lng",
            "required": true
          },
          {
            "schema": {
              "minimum": 1,
              "maximum": 500,
              "default": 500,
              "type": "number"
            },
            "in": "query",
            "name": "radiusM",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/places": {
      "get": {
        "summary": "카카오 장소 검색. 키가 없으면 등록 지점 검색",
        "parameters": [
          {
            "schema": {
              "minLength": 2,
              "maxLength": 100,
              "type": "string"
            },
            "in": "query",
            "name": "q",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/map/items": {
      "get": {
        "summary": "공사·야간 시설·주민 제보 지도 데이터",
        "parameters": [
          {
            "schema": {
              "maxLength": 150,
              "type": "string"
            },
            "in": "query",
            "name": "bbox",
            "required": false
          },
          {
            "schema": {
              "maxLength": 150,
              "type": "string"
            },
            "in": "query",
            "name": "types",
            "required": false
          },
          {
            "schema": {
              "minimum": 0,
              "maximum": 20000,
              "default": 0,
              "type": "integer"
            },
            "in": "query",
            "name": "offset",
            "required": true
          },
          {
            "schema": {
              "minimum": 1,
              "maximum": 1000,
              "default": 500,
              "type": "integer"
            },
            "in": "query",
            "name": "limit",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/v1/facilities/nearby": {
      "get": {
        "summary": "최대 500m 시설·직선거리",
        "parameters": [
          {
            "schema": {
              "minimum": -90,
              "maximum": 90,
              "type": "number"
            },
            "in": "query",
            "name": "lat",
            "required": true
          },
          {
            "schema": {
              "minimum": -180,
              "maximum": 180,
              "type": "number"
            },
            "in": "query",
            "name": "lng",
            "required": true
          },
          {
            "schema": {
              "minimum": 1,
              "maximum": 500,
              "default": 500,
              "type": "number"
            },
            "in": "query",
            "name": "radiusM",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/night/check": {
      "post": {
        "summary": "반딧불이 근접 알림. 위치를 저장하지 않음",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "required": [
                  "position",
                  "accuracyM",
                  "measuredAt",
                  "recentAlerts"
                ],
                "properties": {
                  "position": {
                    "additionalProperties": false,
                    "type": "object",
                    "required": [
                      "lat",
                      "lng"
                    ],
                    "properties": {
                      "lat": {
                        "minimum": -90,
                        "maximum": 90,
                        "type": "number"
                      },
                      "lng": {
                        "minimum": -180,
                        "maximum": 180,
                        "type": "number"
                      }
                    }
                  },
                  "accuracyM": {
                    "minimum": 0,
                    "maximum": 10000,
                    "type": "number"
                  },
                  "measuredAt": {
                    "format": "date-time",
                    "type": "string"
                  },
                  "recentAlerts": {
                    "maxItems": 100,
                    "default": [],
                    "type": "array",
                    "items": {
                      "additionalProperties": false,
                      "type": "object",
                      "required": [
                        "facilityId",
                        "alertedAt"
                      ],
                      "properties": {
                        "facilityId": {
                          "minLength": 1,
                          "maxLength": 100,
                          "pattern": "^[a-zA-Z0-9_-]+$",
                          "type": "string"
                        },
                        "alertedAt": {
                          "format": "date-time",
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/reports": {
      "get": {
        "summary": "공개 제보 목록",
        "parameters": [
          {
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "CONSTRUCTION"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "POOR_LIGHTING"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "FACILITY_FAILURE"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "OBSTACLE"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "ROAD_DAMAGE"
                  ]
                },
                {
                  "type": "string",
                  "enum": [
                    "DANGER"
                  ]
                }
              ]
            },
            "in": "query",
            "name": "type",
            "required": false
          },
          {
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "integer"
            },
            "in": "query",
            "name": "limit",
            "required": true
          },
          {
            "schema": {
              "minimum": 0,
              "maximum": 100000,
              "default": 0,
              "type": "integer"
            },
            "in": "query",
            "name": "offset",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "post": {
        "summary": "현장 제보 등록. 미확인 제보는 경로를 차단하지 않음",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "required": [
                  "title",
                  "description",
                  "position",
                  "type"
                ],
                "properties": {
                  "title": {
                    "minLength": 2,
                    "maxLength": 100,
                    "type": "string"
                  },
                  "description": {
                    "minLength": 10,
                    "maxLength": 2000,
                    "type": "string"
                  },
                  "position": {
                    "additionalProperties": false,
                    "type": "object",
                    "required": [
                      "lat",
                      "lng"
                    ],
                    "properties": {
                      "lat": {
                        "minimum": -90,
                        "maximum": 90,
                        "type": "number"
                      },
                      "lng": {
                        "minimum": -180,
                        "maximum": 180,
                        "type": "number"
                      }
                    }
                  },
                  "type": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "CONSTRUCTION"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "POOR_LIGHTING"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "FACILITY_FAILURE"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "OBSTACLE"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "ROAD_DAMAGE"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "DANGER"
                        ]
                      }
                    ]
                  },
                  "photos": {
                    "maxItems": 2,
                    "type": "array",
                    "items": {
                      "maxLength": 700000,
                      "pattern": "^data:image/(jpeg|png);base64,[A-Za-z0-9+/]+={0,2}$",
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "format": "uuid",
              "type": "string"
            },
            "in": "header",
            "name": "idempotency-key",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/reports/{id}": {
      "get": {
        "summary": "제보 상세",
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "maxLength": 100,
              "pattern": "^[a-zA-Z0-9_-]+$",
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/v1/reports/{id}/flags": {
      "post": {
        "summary": "부적절한 제보 신고",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "required": [
                  "reason"
                ],
                "properties": {
                  "reason": {
                    "minLength": 2,
                    "maxLength": 500,
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "maxLength": 100,
              "pattern": "^[a-zA-Z0-9_-]+$",
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/admin/report-flags": {
      "get": {
        "summary": "운영자 제보 신고 목록",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/reports/{id}/confirmation": {
      "put": {
        "summary": "나도 확인했어요",
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "maxLength": 100,
              "pattern": "^[a-zA-Z0-9_-]+$",
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "delete": {
        "summary": "주민 확인 취소",
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "maxLength": 100,
              "pattern": "^[a-zA-Z0-9_-]+$",
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/v1/admin/reports": {
      "get": {
        "summary": "운영자 전체 제보 목록",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/admin/audit": {
      "get": {
        "summary": "운영 조치 이력 최근 100건",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/me/reports": {
      "get": {
        "summary": "내 제보",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  }
}